requirements-composer.txt
# BioPipelines Workflow Composer Requirements
# ===========================================
#
# Core requirements for the AI Workflow Composer
#
# Install with: pip install -r requirements-composer.txt

# Configuration
pyyaml>=6.0

# Fuzzy string matching for tool search
rapidfuzz>=3.0.0

# LLM Adapters
# ------------

# OpenAI API
openai>=1.0.0

# Anthropic API
anthropic>=0.18.0

# HuggingFace (optional - for local models)
# Uncomment if using local HuggingFace models:
# transformers>=4.30.0
# torch>=2.0.0
# accelerate>=0.20.0

# HTTP requests (for Ollama and other APIs)
requests>=2.28.0
httpx>=0.24.0

# Visualization (optional)
# Uncomment for DAG visualization:
# graphviz>=0.20

# CLI
# ---
rich>=13.0.0  # Rich terminal output

# Data handling
# -------------
pandas>=1.5.0
numpy>=1.24.0

# Testing
# -------
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
# -----------
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0
